refactor: remove deprecated revoke-all-users-improved script and update revoke-all-users script
